removed MosaicConfigFactory
