+++ src/NewChain/Initialization.ts
@@ -6,7 +6,6 @@ import * as ip from 'ip';
 import InitConfig from '../Config/InitConfig';
 import MosaicConfig, { AuxiliaryChain } from '../Config/MosaicConfig';
 import PublishMosaicConfig from "../Config/PublishMosaicConfig";
-import MosaicConfigFactory from '../Config/MosaicConfigFactory';
 import OriginChainInteract from './OriginChainInteract';
 import AuxiliaryChainInteract from './AuxiliaryChainInteract';
 import NodeDescription from '../Node/NodeDescription';
@@ -57,7 +56,7 @@ export default class Initialization {
       auxiliaryNodeDescription,
     );
 
-    const mosaicConfig = MosaicConfigFactory.from(originChainId);
+    const mosaicConfig = MosaicConfig.fromChain(originChainId);
     mosaicConfig.originChain.chain = originChainId;
     mosaicConfig.originChain.contractAddresses.simpleTokenAddress = initConfig.originOstAddress;
